requirements.txt
# Core Dependencies
flask>=3.0.0,<4.0.0
flask-cors>=4.0.0,<5.0.0
flask-limiter>=3.5.0,<4.0.0

# PDF Processing
PyPDF2>=3.0.1,<4.0.0
PyMuPDF>=1.24.0,<2.0.0
pdfplumber>=0.11.0,<1.0.0
pdfminer.six>=20231228
pypdfium2>=4.30.0,<5.0.0
pdf2docx>=0.5.8,<0.6.0
tabula-py>=2.9.0,<3.0.0
pandas>=2.0.0,<3.0.0


# Word Processing
python-docx>=1.1.0,<2.0.0
docx2pdf>=0.1.8,<1.0.0

# Image Processing
pillow>=11.0.0,<12.0.0

# Utilities
requests>=2.31.0
lxml>=5.3.0,<6.0.0
cffi>=1.17.0,<2.0.0
cryptography>=43.0.0,<44.0.0
charset-normalizer>=3.4.0,<4.0.0
pycparser>=2.22,<3.0.0
typing_extensions>=4.12.0,<5.0.0

# Deployment & OS Specific
gunicorn>=21.2.0
pywin32>=308; platform_system == 'Windows'
openpyxl>=3.1.5
python-pptx>=1.0.2
groq>=0.11.0
python-dotenv>=1.0.1
